[SOT][3.11, 3.12, 3.13, 3.14] Refactor Cpython internals
